+++ anchor/programs/bless-token/src/states/bitvec.rs
@@ -1,3 +1,4 @@
+#![allow(dead_code)]
 use anchor_lang::prelude::*;
 
 use crate::errors::BlsError;
