+++ anchor/programs/bless-token/src/states/mod.rs
@@ -1,29 +1,5 @@
 use anchor_lang::prelude::*;
-
-use crate::BITMAP_SIZE;
-
 mod bitvec;
-pub use bitvec::BitVec;
-
-/// RuleOutcome is the fund rule and the fund result.
-#[derive(AnchorDeserialize, AnchorSerialize, Clone, Debug, InitSpace)]
-pub struct RuleOutcome {
-    // the fund rule per month.
-    pub rule: BitVec,
-    // the fund result per month.
-    pub outcome: BitVec,
-}
-
-impl Default for RuleOutcome {
-    fn default() -> Self {
-        Self {
-            // Each account follows a funding schedule spanning 86 months.
-            rule: BitVec::new(BITMAP_SIZE),
-            // Each account tracks funding execution results for 86 months.
-            outcome: BitVec::new(BITMAP_SIZE),
-        }
-    }
-}
 
 /// BlessTokenState is store the bless mint and the packet token accounts.
 /// mint_pubkey is the public key of $BLESS token mint.
@@ -35,62 +11,23 @@ pub struct BlessTokenState {
 
     // gets 16.25%
     pub wallet_investor: Pubkey,
+    pub wallet_investor_token: u64,
 
     //gets 18%
     pub wallet_team_advisor: Pubkey,
+    pub wallet_team_advisor_token: u64,
 
     //gets 10%
     pub wallet_foundation: Pubkey,
+    pub wallet_foundation_token: u64,
 
     //get 10.75%
     pub wallet_ecosystem_liquidityprovision_tgtmarketing: Pubkey,
+    pub wallet_ecosystem_liquidityprovision_tgtmarketing_token: u64,
 
     //gets 45%
     pub wallet_community_rewards: Pubkey,
-
-    pub preseed_sale: Pubkey,
-
-    pub preseed_sale_rule_outcome: RuleOutcome,
-
-    pub seed_sale: Pubkey,
-
-    pub seed_sale_rule_outcome: RuleOutcome,
-
-    pub advisors: Pubkey,
-
-    pub advisors_rule_outcome: RuleOutcome,
-
-    pub team: Pubkey,
-
-    pub team_rule_outcome: RuleOutcome,
-
-    pub ecosystem: Pubkey,
-
-    pub ecosystem_rule_outcome: RuleOutcome,
-
-    pub foundation: Pubkey,
-
-    pub foundation_rule_outcome: RuleOutcome,
-
-    pub market_making: Pubkey,
-
-    pub market_making_flag: bool,
-
-    pub tgt_marketing: Pubkey,
-
-    pub tgt_marketing_flag: bool,
-
-    pub community_airdrop: Pubkey,
-
-    pub community_airdrop_flag: bool,
-
-    pub community_incentives: Pubkey,
-
-    pub community_incentives_flag: bool,
-
-    pub current_month: u8,
+    pub wallet_community_rewards_token: u64,
 
     pub bump: u8,
-
-    pub vault_bump: u8,
 }
